Rewrite ResetCourseAction using ResetTaskFile
There was a problem that we stored text of last submission in TaskFile#text, so reset was impossible.
To fix it now we change only document text of a task file. And because of this it was needed to move
course update from EduStepicConnector#getCourseFromStepik to StudyProjectGenerator#generateProject, as
only there we have all course files and can modify document texts.